chore: remove deleted NeoForge files from old paths
These files were deleted from the working tree (removed in NeoForge
fork) but still tracked. Cleaning up after the repo split.

+++ neoforge-source/src/main/java/net/neoforged/neoforge/event/entity/living/LivingFallEvent.java
@@ -14,32 +14,34 @@ import net.neoforged.neoforge.common.NeoForge;
 /**
  * LivingFallEvent is fired when an Entity is set to be falling.<br>
  * This event is fired whenever an Entity is set to fall in
- * {@link LivingEntity#causeFallDamage(double, float, DamageSource)}.<br>
+ * {@link LivingEntity#causeFallDamage(float, float, DamageSource)}.<br>
  * <br>
- * This event is fired via the {@link CommonHooks#onLivingFall(LivingEntity, double, float)}.<br>
+ * This event is fired via the {@link CommonHooks#onLivingFall(LivingEntity, float, float)}.<br>
  * <br>
  * {@link #distance} contains the distance the Entity is to fall. If this event is canceled, this value is set to 0.0F.
  * <br>
  * This event is {@link net.neoforged.bus.api.ICancellableEvent}.<br>
  * If this event is canceled, the Entity does not fall.<br>
  * <br>
+ * This event does not have a result. {@link HasResult}<br>
+ * <br>
  * This event is fired on the {@link NeoForge#EVENT_BUS}.
  **/
 public class LivingFallEvent extends LivingEvent implements ICancellableEvent {
-    private double distance;
+    private float distance;
     private float damageMultiplier;
 
-    public LivingFallEvent(LivingEntity entity, double distance, float damageMultiplier) {
+    public LivingFallEvent(LivingEntity entity, float distance, float damageMultiplier) {
         super(entity);
         this.setDistance(distance);
         this.setDamageMultiplier(damageMultiplier);
     }
 
-    public double getDistance() {
+    public float getDistance() {
         return distance;
     }
 
-    public void setDistance(double distance) {
+    public void setDistance(float distance) {
         this.distance = distance;
     }
